% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppCompFunk.r
\name{backup}
\alias{backup}
\title{A function primarily for internal use that performs a default backup of COM_LANDS and COMP_EXPAND_DOCS}
\usage{
backup(con, dbName, force = F)
}
\arguments{
\item{con}{An RJDBC connection to an SQL database}

\item{dbName}{A table name to backup, from the given database, given as a string.}

\item{force}{A boolean indicating if a backup should be overwrite an existing
backup made on the same day. By default this function only makes one
backup a day, unless run with force=True to force an overwrite of the
current day's backup.}
}
\description{
A function primarily for internal use that performs a default backup of COM_LANDS and COMP_EXPAND_DOCS
}
