% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageCompFunk.r
\name{estAgeCompDoc}
\alias{estAgeCompDoc}
\title{A version of the core expansion function executing an automated age expansion
based on the borrowing dictated by a doc file.}
\usage{
estAgeCompDoc(
  calcomAgeData,
  doc = sprintf("agedoc\%s.csv", unique(calcomAgeData$tempAge1$year)),
  files = T
)
}
\arguments{
\item{calcomAgeData}{A list as returned by getCalcomAgeData.}

\item{doc}{The filename (or vector of filenames) of the doc files
to use in manual borrowing overrides. Defaults to sprintf("lendoc\%s.csv", unique(calcomLenData$tempLen1$year)}

\item{files}{A boolean flag to produce verbose error files and/or expansion output
files such as sppdoc.}
}
\value{
a data.frame reminiscent of the calcom ___ table.
}
\description{
A version of the core expansion function executing an automated age expansion
based on the borrowing dictated by a doc file.
}
