% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lenCompFunk.r
\name{estLenComp}
\alias{estLenComp}
\title{The core expansion function executing an automated length expansion of Don's
Visual Basic code.}
\usage{
estLenComp(calcomLenData, portBorr = portMatrix1, files = T)
}
\arguments{
\item{calcomLenData}{A list as returned by getCalcomLenData.}

\item{portBorr}{A matrix to define the priority of port borrowing.
Rownames should indicate the port complex code of the actual stratum to
be filled. The first column contains the first priorty for borrowing
from, second column contains the second priority, ... etc. Elements
should be port complex codes as they appear in the samples. Elements not
given as appearing in the samples will code for Nominal.}

\item{files}{A boolean flag to produce verbose error files and/or
expansion output files such as sppdoc.}
}
\value{
a data.frame reminiscent of the calcom ___ table.
}
\description{
The core expansion function executing an automated length expansion of Don's
Visual Basic code.
}
