% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppCompFunk.r
\name{estSppCompDoc}
\alias{estSppCompDoc}
\title{A version of the core expansion function executing an automated species expansion
based on the borrowing dictated by a doc file.}
\usage{
estSppCompDoc(
  pacfinData,
  calcomData,
  doc = sprintf("sppdoc\%s.csv", unique(pacfinData$YR)),
  qtrBorr = qtrMatrix,
  files = T
)
}
\arguments{
\item{pacfinData}{A data.frame as returned by getPacfinSppData.}

\item{calcomData}{A list as returned by getCalcomSppData.}

\item{doc}{The filename (or vector of filenames) of the doc files
to use in manual borrowing overrides. Defaults to sprintf("sppdoc\%s.csv", unique(pacfinData$YR))}

\item{qtrBorr}{A matrix to define the priority of qtr borrowing. Rownames
should define the actual qtr of the stratum to be filled. The first
column contains the the first priority to borrowing from, second column
contains the second priority, ... etc. Elements should be integers \%in\% 1:4.
Elements not given \%in\% 1:4 will code for Nominal.}

\item{files}{A boolean flag to produce verbose error files and/or expansion output
files such as sppdoc.}
}
\value{
a data.frame reminiscent of the calcom.com_lands table.
}
\description{
A version of the core expansion function executing an automated species expansion
based on the borrowing dictated by a doc file.
}
