% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppCompFunk.r
\name{expDiff}
\alias{expDiff}
\title{A function primarily for internal use that compares two expansion objects.}
\usage{
expDiff(exp1, exp2, by = NULL)
}
\arguments{
\item{exp1}{An object that looks like an expansion.}

\item{exp2}{An object that looks like an expansion.}

\item{by}{Which columns to join on.}
}
\description{
A function primarily for internal use that compares two expansion objects.
}
