% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lenCompFuncs.r
\name{exportLen}
\alias{exportLen}
\title{A function to perform the various exporting tasks needed after a species expansion}
\usage{
exportLen(exp, human = T, calcom = F, doc = NULL)
}
\arguments{
\item{exp}{An expansion object created by estLenComp or estLenCompDoc to
be exported.}

\item{human}{A boolean indicating if the human feed should be written to
file (i.e. hfeedYY.csv).}

\item{calcom}{A boolean indicating if the the calcom database should be
backed up and updated with the given species expansion.}

\item{doc}{If calcom=True, a vector of doc filenames to be exported to
calcom.}
}
\value{
NULL See current directory and/or database.
}
\description{
A function to perform the various exporting tasks needed after a species expansion
}
