% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppCompFunk.r
\name{exportSpp}
\alias{exportSpp}
\title{A function to perform the various exporting tasks needed after a species expansion}
\usage{
exportSpp(exp, human = T, pacfin = T, calcom = F, doc = NULL)
}
\arguments{
\item{exp}{An expansion object created by estSppComp or estSppCompDoc to
be exported.}

\item{human}{A boolean indicating if the human feed should be written to
file (i.e. hfeedYY.csv).}

\item{pacfin}{A boolean indicating if the pacfin feed should be written to
file (i.e. pfeedYY.DAT).}

\item{calcom}{A boolean indicating if the the calcom database should be
backed up and updated with the given species expansion.}

\item{doc}{If calcom=True, a vector of doc filenames to be exported to
calcom.}
}
\value{
NULL See current directory and/or database.
}
\description{
A function to perform the various exporting tasks needed after a species expansion
}
