% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppCompFunk.r
\name{getCalcomSppData}
\alias{getCalcomSppData}
\title{Collect and prepare Calcom data for species expansion.}
\usage{
getCalcomSppData(year, save = F, fromFile = F)
}
\arguments{
\item{year}{The year of the expansion given as a four digit integer. Year may
be given as a vector to prepare data for a multi-year expansion.}

\item{save}{A filename.csv to save data to file or a logical. If save=True
the filename defalts to sprintf('calcom\%sSppData\%s', year, Sys.time());
save=False (Default) does not save.}

\item{fromFile}{A filename.csv to read data from file or a logical.
fromFile=True defaults to the most recent sprintf('pacfin\%sData*', year)
file; fromFile=False (Default) reads data from calcom.}
}
\value{
Returns a list of the various objects from calcom needed to compute
a species expansions for the given years.
}
\description{
Collect and prepare Calcom data for species expansion.
}
