% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppCompFunk.r
\name{getDrivers}
\alias{getDrivers}
\title{A function primarily for internal use that checks to make sure the database drivers are in place. Requires an internet connection to download.}
\usage{
getDrivers()
}
\description{
A function primarily for internal use that checks to make sure the database drivers are in place. Requires an internet connection to download.
}
