% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppCompFuncs.r
\name{getPacfinSppData}
\alias{getPacfinSppData}
\title{Collect and prepare Pacfin data for species expansion.}
\usage{
getPacfinSppData(year, save = F, fromFile = F)
}
\arguments{
\item{year}{The year of the expansion given as landing_year is expressed in
the pacfin_marts.comprehensive_ft table. Year may be given as a vector to
prepare data for a multi-year expansion.}

\item{save}{A filename.csv to save data to file or a logical. If save=True
the filename defalts to sprintf('pacfin\%sSppData\%s', year, Sys.time());
save=False (Default) does not save.}

\item{fromFile}{A filename.csv to read data from file or a logical.
fromFile=True defaults to the most recent sprintf('pacfin\%sSppData*', year)
file; fromFile=False (Default) reads data from pacfin.}
}
\value{
Returns a data.frame of the neccessary fish ticket data from pacfin
to perform species expansions for the given years
}
\description{
Collect and prepare Pacfin data for species expansion.
}
