% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppCompFunk.r
\name{getSelfPath}
\alias{getSelfPath}
\title{A function primarily for internal use that finds path of the package install. Built for use in referencing the driver location.}
\usage{
getSelfPath()
}
\description{
A function primarily for internal use that finds path of the package install. Built for use in referencing the driver location.
}
