% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lenCompFunk.r
\name{howFar}
\alias{howFar}
\title{A function primarily for internal use that returns the number of port complex away a given target port is from a hypothetical borrowed port.}
\usage{
howFar(tar, bor)
}
\arguments{
\item{tar}{target port}

\item{bor}{borrowed port}
}
\description{
A function primarily for internal use that returns the number of port complex away a given target port is from a hypothetical borrowed port.
}
