% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppCompFunk.r
\name{makeCodeStrat}
\alias{makeCodeStrat}
\title{A function primarily for internal use that decodes the strata from Don's magic number coding.}
\usage{
makeCodeStrat(code)
}
\arguments{
\item{code}{Don's magic number}
}
\description{
A function primarily for internal use that decodes the strata from Don's magic number coding.
}
