% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppCompFuncs.r
\name{makeStratCode}
\alias{makeStratCode}
\title{A function primarily for internal use that encodes a strata as Don's magic number coding.}
\usage{
makeStratCode(disp, port, gear, mcat, qtr)
}
\arguments{
\item{disp}{disposition}

\item{port}{three letter port code}

\item{gear}{three letter gear code}

\item{mcat}{numeric market category}

\item{qtr}{numeric quarter}
}
\description{
A function primarily for internal use that encodes a strata as Don's magic number coding.
}
