% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppCompFunk.r
\name{pacfinPortCode}
\alias{pacfinPortCode}
\title{A function primarily for internal use that converts three letter port codes to pacfin port codes.}
\usage{
pacfinPortCode(port)
}
\arguments{
\item{port}{three letter port code}
}
\description{
A function primarily for internal use that converts three letter port codes to pacfin port codes.
}
