% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppCompFuncs.r
\docType{data}
\name{portMatrix2}
\alias{portMatrix2}
\title{A default matrix that encodes the default 2-away port sharing rules (no borrowing across Point Conception).}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 10 rows and 5 columns.
}
\usage{
portMatrix2
}
\description{
Row index defines the port to be filled in north to south encoding
c('CRS', 'ERK', 'BRG', 'BDG', 'OSF', 'MNT', 'MRO', 'OSB', 'OLA', 'OSD').
Column values define the priority of sharing across ports for the given row.
Any value outside of c('CRS', 'ERK', 'BRG', 'BDG', 'OSF', 'MNT', 'MRO', 'OSB', 'OLA', 'OSD') codes for 'NOMINAL'.
}
\keyword{datasets}
