% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppCompFunk.r
\docType{data}
\name{qtrMatrix}
\alias{qtrMatrix}
\title{A default matrix that encodes the default qtr sharing rules from Dons code.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 3 columns.
}
\usage{
qtrMatrix
}
\description{
Row index defines the qtr to be filled and column values define the priority of sharing across qtrs for the given row.
}
\keyword{datasets}
